Moved services into course module and added providers in the module (#194)
