+++ src/app/problems/problem-create/parsons-create-snippet/parsons-create-snippet.component.ts
@@ -4,7 +4,7 @@ import {Category, Course} from '@app/_models';
 import {CourseEvent} from '@app/_models/course_event';
 import {QuestionService} from '@app/problems/_services/question.service';
 import {ToastrService} from "ngx-toastr";
-import {CourseService} from '@app/_services/api/course/course.service';
+import {CourseService} from '@app/course/_services/course.service';
 import {CategoryService} from '@app/_services/api/category.service';
 import {forkJoin} from 'rxjs';
 import {ParsonsForm} from "@app/problems/_forms/parsons.form";
