+++ src/app/problems/problem-edit/parsons-edit-snippet/parsons-edit-snippet.component.ts
@@ -5,9 +5,9 @@ import {ToastrService} from "ngx-toastr";
 import {Category, Course} from '@app/_models';
 import {CourseEvent} from '@app/_models/course_event';
 import {forkJoin} from 'rxjs';
-import {CourseService} from '@app/_services/api/course/course.service';
+import {CourseService} from '@app/course/_services/course.service';
 import {CategoryService} from '@app/_services/api/category.service';
-import {CourseEventService} from '@app/_services/api/course/course-event.service';
+import {CourseEventService} from '@app/course/_services/course-event.service';
 import {ParsonsForm} from "@app/problems/_forms/parsons.form";
 import {Router} from "@angular/router";
